Add files via upload
Animation is a method of photographing successive drawings, models, or even puppets, to create an illusion of movement in a sequence. Because our eyes can only retain an image for approximately 1/10 of a second, when multiple images appear in fast succession, the brain blends them into a single moving image. 

In traditional animation, pictures are drawn or painted on transparent celluloid sheets to be photographed. Early cartoons are examples of this, but today, most animated movies are made with computer-generated imagery or CGI.

To create the appearance of smooth motion from these drawn, painted, or computer-generated images, frame rate, or the number of consecutive images that are displayed each second, is considered. Moving characters are usually shot “on twos” which just means one image is shown for two frames, totaling in at 12 drawings per second. 12 frames per second allows for motion but may look choppy. In the film, a frame rate of 24 frames per second is often used for smooth motion.